+++ CHANGELOG.md
@@ -6,6 +6,9 @@
 - build: disable AndroidStripILAfterAOT
 - BouncyCastle.Cryptography 2.6.2
 - Microsoft.Extensions.Logging.Abstractions 9.0.8
+- Square.OkHttp3.JVM 5.1.0.2 (replaces Square.OkHttp3)
+- Square.OkHttp3.JavaNetCookieJar 5.1.0.2 (replaces Square.OkHttp3.UrlConnection)
+- Square.OkIO 3.16.0
 - test : Microsoft.Extensions.DependencyInjection 9.0.8
 - test : Microsoft.Extensions.Logging 9.0.8
 - test : Microsoft.Maui.* 9.0.100